requirements.txt
pandas==2.2.3
openpyxl==3.1.5
openai==1.61.1
google-genai=1.0.0
Pillow=11.1.0
google-generativeai=0.8.4
pymilvus=2.5.4
tqdm=4.67.1
ipywidgets=8.1.5
bs4=0.0.2
transformers=4.48.3
streamlit=1.42.0
watchdog=6.0.0
torch=2.6.0
FlagEmbedding=1.3.4